requirements.txt
streamlit
pypdf
markdown
sentence-transformers
faiss-cpu
transformers
torch
